+++ ConsoleExample/Attributes/AdminOnlyExampleAttribute.cs
@@ -1,0 +1,6 @@
+﻿namespace ConsoleExample.Attributes
+{
+    internal class AdminOnlyExampleAttribute : Attribute
+    {
+    }
+}