+++ PRTelegramBot/Core/UpdateHandlers/CommandsUpdateHandlers/InlineUpdateHandler.cs
@@ -35,10 +35,13 @@ namespace PRTelegramBot.Core.UpdateHandlers.CommandsUpdateHandlers
                     string msg = $"The user {update.GetInfoUser().Trim()} invoked the command {command.CommandType.GetDescription()}";
                     bot.Events.OnCommonLogInvoke(msg, "CallBackCommand", ConsoleColor.Magenta);
 
+                    bot.Events.MessageEvents.OnPreInlineCommandHandleInvoke(new BotEventArgs(bot, update));
                     var resultExecute = await ExecuteCommand(command.CommandType, update, commands);
                     if (resultExecute == CommandResult.Executed)
+                    {
+                        bot.Events.MessageEvents.OnPostInlineCommandHandleInvoke(new BotEventArgs(bot, update));
                         return UpdateResult.Handled;
-
+                    }
                 }
                 return UpdateResult.Continue;
             }